+++ .github/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [4.7.4] - 2021-06-05
+*No changes, CI test release*
+
 ## [4.7.3] - 2021-06-05
 ### Fixed
 * Forced verifications no longer strip the first word off the message
@@ -400,7 +403,8 @@ Now it compiles!
 ## [1.0.0] - 2020-11-30
 * Initial release with `bio` component
 
-[Unreleased]: https://github.com/codemicro/lgballtDiscordBot/compare/v4.7.3...HEAD
+[Unreleased]: https://github.com/codemicro/lgballtDiscordBot/compare/v4.7.4...HEAD
+[4.7.4]: https://github.com/codemicro/lgballtDiscordBot/compare/v4.7.3...v4.7.4
 [4.7.3]: https://github.com/codemicro/lgballtDiscordBot/compare/v4.7.2...v4.7.3
 [4.7.2]: https://github.com/codemicro/lgballtDiscordBot/compare/v4.7.1...v4.7.2
 [4.7.1]: https://github.com/codemicro/lgballtDiscordBot/compare/v4.7.0...v4.7.1